+++ src/Molecules/FlexTable/Cell/Cell.tsx
@@ -6,7 +6,7 @@ import { Flexbox } from '../../..';
 import { useColumnLayout } from '../shared/ColumnProvider';
 import { CellComponent } from './Cell.types';
 import { TextWrapper } from './TextWrapper';
-import { TruncateWithTooltip } from '../shared';
+import { useFlexTable } from '../shared/FlexTableProvider';
 
 const StyledFlexbox = styled(Flexbox)`
   overflow: hidden;
@@ -14,8 +14,9 @@ const StyledFlexbox = styled(Flexbox)`
 
 // TODO: Fix typings that react memo causes when exporting textwrapper on cell
 // @ts-ignore
-const Cell: CellComponent = React.memo(({ children, className, fontSize, columnId }) => {
+const Cell: CellComponent = React.memo(({ children, className, columnId }) => {
   const [columnLayout] = useColumnLayout(columnId);
+  const { fontSize } = useFlexTable();
 
   if (!R.prop('flexProps', columnLayout)) {
     return null;
@@ -26,11 +27,7 @@ const Cell: CellComponent = React.memo(({ children, className, fontSize, columnI
       {isElement(children) && children}
       {isFunction(children)
         ? children({ fontSize, columnId })
-        : !isElement(children) && (
-            <TruncateWithTooltip label={children} position="top">
-              <TextWrapper fontSize={fontSize}>{children}</TextWrapper>
-            </TruncateWithTooltip>
-          )}
+        : !isElement(children) && <TextWrapper fontSize={fontSize}>{children}</TextWrapper>}
     </StyledFlexbox>
   );
 });