+++ src/Molecules/FlexTable/FlexTable.tsx
@@ -8,7 +8,7 @@ import { constants, ColumnProvider } from './shared';
 import { Props, FlexTableComponents } from './FlexTable.types';
 import { FlexTableProvider, useFlexTable } from './shared/FlexTableProvider';
 import { ExpandCell } from './Cell/ExpandCell';
-import { Typography, Box } from '../..';
+import { Typography } from '../..';
 import { isElement } from '../../common/utils';
 
 type HtmlDivProps = {} & React.HTMLAttributes<HTMLDivElement>;
@@ -33,26 +33,35 @@ const FlexTableContainer: React.FC<HtmlDivProps> = ({ className, children, ...ht
   );
 };
 
+const StyledCaption = styled.caption`
+  display: flex;
+`;
+
+const StyledTypography = styled(Typography)`
+  padding-left: ${p => p.theme.spacing.unit(1)}px;
+`;
+
 const FlexTable: React.FC<Props> & FlexTableComponents = ({
   className,
   density = 'm',
+  expandable = false,
   stickyHeader = true,
   children,
   title,
+  fontSize = 'm',
   ...htmlProps
 }) => (
-  <FlexTableProvider density={density} stickyHeader={stickyHeader}>
+  <FlexTableProvider
+    density={density}
+    stickyHeader={stickyHeader}
+    fontSize={fontSize}
+    expandable={expandable}
+  >
     {/* pass sticky with context instead of prop-drilling, since context might change */}
     <FlexTableContainer className={className} {...htmlProps}>
-      <caption>
-        {isElement(title) ? (
-          title
-        ) : (
-          <Box pl={1}>
-            <Typography type="title3">{title}</Typography>
-          </Box>
-        )}
-      </caption>
+      <StyledCaption>
+        {isElement(title) ? title : <StyledTypography type="title3">{title}</StyledTypography>}
+      </StyledCaption>
       <ColumnProvider>{children}</ColumnProvider>
     </FlexTableContainer>
   </FlexTableProvider>