+++ src/Molecules/FlexTable/Header/Header.tsx
@@ -15,7 +15,6 @@ import {
   useColumnLayout,
 } from '../shared/ColumnProvider';
 import { HeaderContent, TextWrapper, SortIcon, SortButton } from './HeaderContent';
-import { TruncateWithTooltip } from '../shared';
 
 const SORT_ORDERS: SortOrder[] = [SORT_ORDER_NONE, SORT_ORDER_ASCENDING, SORT_ORDER_DESCENDING];
 
@@ -43,7 +42,6 @@ const Header: HeaderComponent = props => {
     sortable = false,
     sortOrder: sortOrderProp,
     onSort = () => {},
-    fontSize,
     columnId,
   } = props;
 
@@ -96,19 +94,16 @@ const Header: HeaderComponent = props => {
     >
       {isElement(children) && children}
       {isFunction(children)
-        ? children({ fontSize, sortable, sortOrder, onSortClick, sorted, columnId })
+        ? children({ sortable, sortOrder, onSortClick, sorted, columnId })
         : !isElement(children) && (
-            <TruncateWithTooltip position="top" label={children}>
-              <HeaderContent
-                onSortClick={onSortClick}
-                sortable={sortable}
-                sortOrder={sortOrder}
-                fontSize={fontSize}
-                sorted={sorted}
-              >
-                {children}
-              </HeaderContent>
-            </TruncateWithTooltip>
+            <HeaderContent
+              onSortClick={onSortClick}
+              sortable={sortable}
+              sortOrder={sortOrder}
+              sorted={sorted}
+            >
+              {children}
+            </HeaderContent>
           )}
     </StyledFlexbox>
   );