+++ src/Molecules/FlexTable/Header/HeaderContent/HeaderContent.types.ts
@@ -27,4 +27,4 @@ type onSortClick = () => void;
 
 export type UIProps = { onSortClick: onSortClick };
 
-export type Props = { sortable: boolean; sortOrder: SortOrder; sorted: boolean } & TextWrapperProps;
+export type Props = { sortable: boolean; sortOrder: SortOrder; sorted: boolean };