+++ src/Molecules/FlexTable/shared/FlexTableProvider/FlexTableProvider.types.ts
@@ -1,8 +1,10 @@
-import { Density } from '../shared.types';
+import { Density, FontSize } from '../shared.types';
 
 export type FlexTableState = {
   density: Density;
   stickyHeader: boolean;
+  fontSize: FontSize;
+  expandable: boolean;
 };
 
 export type Props = {} & FlexTableState;